+++ ChangeLog.md
@@ -41,6 +41,8 @@ Different versions of mealpy in terms of passing hyper-parameters. So please car
 + Add `Cosmic Evolution Optimization (CEO)` to the `physics-based` group.
 + Add `Electric Eel Foraging Optimization (EEFO)` to the `swarm-based` group.
 + Add `Artificial Protozoa Optimizer (APO)` to the `bio-based` group
++ Add `Artificial Protozoa Optimizer (APO)` to the `bio-based` group
++ Add `General Relativity Search Algorithm (GRSA)` to the `physics-based` group
 
 ---